backend/requirements.txt
asgiref==3.10.0
boto3==1.40.64
botocore==1.40.64
Django==5.2.7
jmespath==1.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
django-storages
boto3